requirements.txt
pyodbc
python-dotenv
argon2-cffi
pillow